added hash_digest to passwords, can only delete if logged in, can only update/create if provides correct password

+++ tep2019_code/djangular/src/app/teacher-detail/teacher-detail.component.ts
@@ -51,6 +51,10 @@ export class TeacherDetailComponent implements OnInit {
   }
 
   deleteTeacher(teacher: TeacherDetail) {
+    if (this.teacher.orders.length > 0) {
+      alert("Cannot delete this teacher. They have an order in the system. Please delete the order first if you wish to preceed");
+      return;
+    }
     this.apiService.delete("teachers", this.teacher.id).subscribe();
     this.location.back();
   }
